make `importPackages("")` and `importClasspath(..)` consistent
So far `ClassFileImporter.importPackages("")` would only return classes returned by `ClassLoader.getResources("")`, but not other classes derived from the configured URL classpath (as done for other prefixes like `com`, which would be searched through the whole classpath as well). The reason is that this corner case has been overlooked in `NormalizedResourceName.startsWith(..)`. I.e. if the prefix is empty, then the resource would only be included if it starts with a `/`, but resources in archives do not start with a `/`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>