add `ImportOption` `DoNotIncludeGradleTestFixtures`
Gradle provides the `java-test-fixtures` plugin out of the box, which will add a separate source set for creating test fixtures (i.e. test support code). This source set can see the main sources and be seen by the test sources. Depending on how the tests are run they might be put onto the classpath as folder path (containing the compiled sources) or JAR archive. Thus, we have to exclude both, the folder where Gradle compiles the files to by convention and archives that follow the Gradle convention to end in `-test-fixtures.jar`.
However, this can of course only be a heuristic. Custom modules that mimic the convention will e.g. also be excluded. But in these cases users have to customize their import options then, instead of using the predefined one.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>