mark classes that are part of the public API as `@PublicAPI`
So far we only marked members that were accessible as `@PublicAPI`, which made classes `@PublicAPI` implicitly. Some classes are also marked as `@PublicAPI` at the class level. We should make this consistent and easy to understand for users, so we mark every class toplevel as `@PublicAPI` that is part of the public API.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>