report classes not contained in any PlantUML component as violation
So far when checking an `ArchRule` based on `PlantUmlArchCondition` we were throwing an exception if a class was not contained in any diagram component. This causes problems for legacy code bases where some classes might live outside the desired components, which should be frozen as violations to be iteratively fixed. But throwing an `IllegalStateException` prevents any such approach. We thus replace the `IllegalStateException` by a regular rule violation that can be treated like any other violation of the architecture and in particular be frozen via `FreezingArchRule`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>