report classes contained in multiple PlantUML components as violation
So far when checking an `ArchRule` based on `PlantUmlArchCondition` we were throwing an exception if a class was contained in multiple diagram components. This causes problems for legacy code bases where some classes might violate the conventions in such a way, which should be frozen as violations to be iteratively fixed. But throwing a `ComponentIntersectionException` prevents any such approach. We thus replace the `ComponentIntersectionException` by a regular rule violation that can be treated like any other violation of the architecture and in particular be frozen via `FreezingArchRule`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>