rename `JavaPackage.getAllClasses()` -> `getClassesInPackageTree()`
In all other places we have replaced the fuzzy word "all" by "in package tree" to clarify the semantics. The Javadoc already hinted at the fact that this method name should be consolidated. Now with release 1.0.0 we can break the API and don't need to deprecate the old method.
I decided to call the method `getClassesInPackageTree()`, even though the similar method `getSubpackagesInTree()` does not add the infix "package", because I think for the former it is worth it to clarify the context, while the latter already talks about packages, so "tree" is already clear from the context.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>