support rc-versions in `publish.gradle`
The latest release version gets parsed from the README.md file. So far it was always expected to consist of the major.minor.patch format without any suffixes. Now that the latest release version is `1.0.0-rc1` this crashes. I extended the parsing to allow suffixes, since this seems to be a valid use case in general.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>