fix referenced class objects not detected in lambdas
With a404fb4b we changed the import of lambdas, such that accesses declared within lambdas are not associated with the synthetic lambda method anymore, but instead with the real method surrounding the lambda. Unfortunately we overlooked non-access dependencies declared within lambdas, like referenced class objects. Since we now filter out the synthetic lambda methods now completely, this meant that those dependencies now completely vanished from the import.
We now fix this by applying the same non-synthetic origin resolution mechanism to `RawReferencedClassObject` that we only applied to `RawAccessRecord` before. This also means, that we can't simply add `RawReferencedClassObject` directly to the `JavaCodeUnitBuilder` anymore and in particular can't resolve it directly on `JavaCodeUnit` creation, but instead supply it later on like the accesses.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>