always compile older sources with a JDK LTS version
We don't want to depend on some EOL JDKs anymore in CI, but right now when we choose JDK 8 for certain modules the logic will say "JDK 8 is too old for this module, it needs JDK 9", then download a JDK 9 through the toolchain to compile. We now make sure we pick the next higher compatible LTS version (e.g. JDK 11 if the `minimumJavaVersion` of the module is JDK 9).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>