make `TextFileBasedViolationStore` public API
At the moment it is quite impossible to make any adjustments to the default `TextFileBasedViolationStore`, since the class is fully hidden from the public API and can thus not be instantiated or modified with respect to the implementation (e.g. adjusting the file names of the store).
We now make it public API and allow users access to it, so it can be extended through delegation.

Signed-off-by: danhaywood <dan@haywood-associates.co.uk>
Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>