improve naming of `RandomSyntaxTestBase` components
The naming was a little confusing, i.e. that a set of `SpecificParameterProvider` would be called `singleParameterProviders`, but then there is also a class `SingleParametersProvider`, etc. The new names should reflect the role a little better. I also moved the `SingleParameterProvider` toplevel, so we can adjust it per test level in the next step.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>