generify synthetic access resolution
We really only need "something that has a code unit origin" to resolve to correct non-synthetic origin, it doesn't necessarily need to be a "code unit dependency". This is relevant, because we want to fix try-catch-blocks as well (i.e. if the try-catch-block is defined within a lambda), and such a try-catch-block doesn't really conform to the concept of a "code unit dependency" (the thrown type, etc., likely would, but that's not what we want to fix).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>