fix missing try-catch-blocks declared in lambda
We need to resolve synthetic origins for try-catch-blocks just like other objects declared in `JavaCodeUnit`, like `JavaAccess` or `InstanceofCheck`. Otherwise, a `TryCatchBlock` declared within a lambda will be imported as being declared within a synthetic method (like `lambda$abc$123`). But we throw out these synthetic methods in the end, because the origins should have been fixed during the import to point to the method declaring the lambda. Thus, before this change we lost all those try-catch-blocks together with these methods.

Resolves: #1069

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>