ensure all `AccessRecord`s are only created once when fixing origin
To speed up performance we often use the identity equals and hashcode for objects like `AccessRecord`. This works i.g. quite well, because most objects are created uniquely only once, so doing a semantic equals (based on the properties) or a identity equals is in fact the same. When we fix the synthetic origins we need to make sure that we also only create each fixed `AccessRecord` once. At the moment, this is actually the case, because the `forEach...` methods that fix the origins are always only called once. But this is for once a very implicit requirement (that these methods should only be called once if we want to compare the `AccessRecord`s to each other), and secondly to fix synthetic accesses in try-catch-blocks we need to ensure that the `AccessRecord`s referenced within `RawTryCatchBlock` are in fact equal again to the `AccessRecord`s referenced outside. The easiest way to keep the identity consistent, and ensure that fixed records are also only created once under all circumstances, seems to be to cache the fixed `AccessRecord`s.

Note, that it really makes a measurable difference for the necessary heap space if we store a singleton set with the original access for every non-synthetic access. Thus, we only cache the set of resolved accesses if a synthetic access has been resolved and use an empty set marker for all regular values.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>