unify URL parsing into base and path
We use different approaches in different places to obtain base, path or both from a JAR URI (i.e. the part up to the separator denoting where the JAR resides versus the part after the separator denoting the path within the JAR file). We now unify this to make the connection clearer in code and use simple character splitting instead of Regex.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>