add convenience API to specify allowed module dependencies by name
This is a simplification that allows to directly write down which modules may depend on which other modules by name. I picked the module name for this definition and not the identifier, because if the two differ I think the name is nicer to read (and also unique as the identifier).

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>