ignore unknown components in dependencies of PlantUML diagrams
The current implementation prevents using PlantUML diagrams as rules
that are used for wider documentation purposes (e.g. containing a
database component or some other dependency that's irrelevant from the
point of an ArchUnit test).
The consequence of ignoring such unknown components is only to allow
fewer dependencies and fewer classes to be present.
So, at best the test will fail as a false positive, if the components
are declared wrongly by accident.
Thus, there is no major harm in simply ignoring these unknown components
and giving users a little more freedom what elements they want to use
in their diagrams.

Issue: #1132

Signed-off-by: Tomasz Fijalkowski <3066563+tfij@users.noreply.github.com>