add up-to-date check for release check
It happened several times that a release dependency (e.g. `slf4j-api`) was updated e.g. by Dependabot, but the respective expected POM template for the release check was forgotten to adjust.
This then makes the release fail, because it thinks the produced POM file doesn't match the expected one.
We now ensure during build that the release check POM file dependency versions still match the project dependencies, so this can't go out of sync anymore.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>