support optional `component` keyword in PlantUML diagram rules
It is valid PlantUML syntax to prepend the `component` keyword to a component
declaration, e.g.

```
component [MyComponent] <<..mycmp..>>
```

Since it doesn't create a big overhead to support this as an optional part
of the component declaration, we can add it to make it easier for users
that distinguish their components from other types of objects like databases.

Signed-off-by: Tomasz Fijalkowski <3066563+tfij@users.noreply.github.com>