run archunit-junit5 integration tests non-deprecated
We had used the JUnit 4 compatible `JUnitPlatform` runner
to run archunit-junit5 examples programmatically for the integration tests.
This class is now deprecated, and it also makes no sense to couple
running the tests to the old JUnit 4 infrastructure.
Thus, we switch this to `LauncherFactory` to run it the recommended way.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>