pass `SourceCodeLocation` on to construct `Dependency`
At the moment the line number is passed to the `Dependency`,
then it constructs a `SourceCodeLocation` from the `originClass` and the `lineNumber`.
In all cases we have a `SourceCodeLocation` outside already though.
And the `SourceCodeLocation` constructed in the constructor of `Dependency`
will just create an equivalent `SourceCodeLocation` to what we already have
when calling the constructor.
We can thus just pass on the `SourceCodeLocation` from the outside.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>