fix wrong modifiers on imported array type class objects
The Reflection API considers the visibility of an array type class object
to be equal to the visibility of the element type class object.
ArchUnit so far considered every array type class object as public.
Since we usually strive for being as similar to the Reflection API as possible,
we fix this now by deriving the visibility from the element type as well.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>