fix `getAllInvolvedRawTypes()` recursing infinitely
For recursive type parameter definitions like `class Example<T extends Comparable<T>>`
the method `JavaClass.getAllInvolvedRawTypes()` causes a `StackOverflowError`
at the moment. This is because we have a simple recursion without any short circuit
condition if we encounter a type we've already seen before.
We can utilize `JavaType.traverseSignature(..)` to solve this,
because there the problem is already solved in the visitor pattern implementation.

This reverts commit 6f23f66fd7f355fe40bea5b5a0fcd18a17e2e783.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>