fix JUnit 5 tests not being executed in CI
When migrating the build configuration to plugins,
we accidentally disabled JUnit 5 tests being executed during CI.
Since we don't have many such tests and nothing ever broke there,
this went unnoticed until now.
The reason is that there is a generic piece
of configuration in `java-testing-conventions.gradle`
that reacts to `archUnitTest.hasSlowTests == true`,
and if `true` and the property `allTests` not set,
configures JUnit 4 to exclude the Category `Slow`.
Unfortunately, this also switches the test execution completely back to JUnit 4,
disabling any JUnit 5 tests to be run.
And since we configure this in `afterEvaluate`,
it actually overwrites the test configuration of the concrete module.
Ironically, we don't even have any `Slow` tests in modules
where we use JUnit 5.
Thus, we can simply disable `archUnitTest.hasSlowTests` for those.
We include a check for now to make sure that this will never be set again
for modules that contain JUnit 5 tests.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>