fix test location for nested `ArchTests.in(..)` with JUnit 4
So far, if we include rules from other test classes, e.g.
```
@RunWith(ArchUnitRunner.class)
@AnalyzeClasses(..)
class ExampleTest {
  @ArchTest
  static final ArchTests nested = ArchTests.in(Other.class);
}
```
then the nested class (`Other.class` in this case) will be used as test location.
This can cause confusion, because by this if `Other.class` is included in two separate test classes,
analyzing different locations via `@AnalyzeClasses`,
the results will be reported for the same test class `Other.class`.
We now set the root class that started the test as the test class of the `Description`.
To make it easy to understand through which path a rule is included in more complex scenarios,
we extend the display name to include the path of nested classes.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>