fix `@ArchIgnore` not being fully respected by JUnit 4 support
In the case of inheritance we were looking at the wrong class for an `@ArchIgnore`.
I.e. rule declarations in base classes would only be ignored if the base class
was annotated with `@ArchIgnore`, not the actual class.
We now check the actual class for the annotation instead.
If the base class is annotated with `@ArchIgnore` we ignore it for now,
since this is a corner case and one could argue for both,
either ignoring all subclasses or not ignoring subclasses.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>