fix test location for nested `ArchTests.in(..)` with JUnit 5
So far, if we include rules from other test classes, e.g.
```
@AnalyzeClasses(..)
class ExampleTest {
  @ArchTest
  static final ArchTests nested = ArchTests.in(Other.class);
}
```
then the nested class (`Other.class` in this case) will be used as test location.
This can cause confusion, because by this if `Other.class` is included in two separate test classes,
analyzing different locations via `@AnalyzeClasses`,
the results will be reported for the same test class `Other.class`.
At least for Gradle the reason it picks `Other.class` seems to be
that it traverses up the `TestDescriptor` hierarchy
and picks the first descriptor with an attached `ClassSource`.
We can thus simply not set a `TestSource` for the intermediary descriptors,
i.e. the container descriptors caused by `ArchTests.in(..)`,
and by that make Gradle traverse up until it finds the root that started the test.
To make it easy to understand through which path a rule is included in more complex scenarios,
we extend the display name to include the path of nested classes.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>