support filtering `@ArchTest` members by system property for JUnit 4
One inconvenience of the current state of the `ArchUnitRunner`
is the fact that there is no simple way to execute a single rule.
For methods annotated with `@ArchTest` this might be supported by the IDE,
but for fields there is no convenient way except of commenting all other `@ArchTest` annotations.

We now provide a simple way to run any member in isolation
by specifying a system property / `ArchConfiguration` value.
This change allows to pass a simple test name filter
that for now only supports exact member matches.
E.g. specifying `-Darchunit.junit.testFilter=my_custom_rule` would run the rule
declared in a field or member with name `my_custom_rule`.
Multiple members can be joined by `,` and are then all executed in one run.

Unfortunately, filtering in JUnit 4 is tied to only the `Description` of the current test.
Since we don't want to couple text parsing of the description to filtering
(e.g. if we change the description in the future),
we have to introduce a hack to transmit the original member name to the `Filter`.
Due to the fact that `Description` is effectively final and has no intended way for our
use case, we (ab-)use the `annotations` of the test member to pass along the information
in a custom annotation. This should have no further effect, since no infrastructure can
react to unknown annotations, which may be present at any time for given tests.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>