update ArchUnit-Examples dependencies on release
We automatically keep the code of the separate ArchUnit-Examples repository
in sync with the `archunit-example` module in this repository.
However, sometimes an example makes an update of the dependencies necessary
(as seen in 1d1a990eea6f7babb29847ad3b42d37c56f81f47).
Up to now this made it necessary to manually update the dependencies in ArchUnit-Examples,
or the release will fail when testing the updated ArchUnit-Examples.
We now keep the (production) dependencies of ArchUnit-Examples automatically in sync
with the `archunit-example` module.

Since ArchUnit-Examples should be purely driven from this repository,
I think it's an okay compromise to simply keep a static template here
and fully overwrite the build file.
Changes to the root build file should always be controlled from this repository.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>