make out of classpath testing more robust
It has always been a little brittle to use classes
that were compiled regularly and reachable from the
default ClassLoader for test "outside the classpath".
The reason is that if such a class is referenced at any point
before the test runs, then the class is already loaded.
And then we can move these classes out of the folder in the file
system later on all we want, the ClassLoader can still load them
(this was also the reason why we had to use strings for the class names
in such tests instead of referencing the `Class` objects).
I found out that switching to JUnit 5 this actually becomes a problem,
because something in the build infrastructure seems to trigger
a class loading of all classes before already with the test ClassLoader
(maybe it's scanning classes for annotations and by that triggering class loading).

To make the test infrastructure more robust we now don't use regularly
compiled classes out of the URL classpath of the ClassLoader,
but instead we keep the source files as resources and compile them
dynamically when needed.
This way we can be sure that the regular test ClassLoader has no chance
to ever know these classes ahead of being used by a test.
And it also has no chance to load these classes anywhere from the
regular classpath.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>