disable updating website & examples
This is just a release for Spring Modulith,
because their update policy prevents updating to higher minor versions like 1.4.0 for the current Modulith release.
But, since 1.4.0 has already been released,
we don't want to update website and examples to show 1.3.1 as the current version,
even if it has been released chronologically after 1.4.0.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>