fix user guide describing wrong default behavior of `@AnalyzeClasses`
The user guide claims that without specifying anything `@AnalyzeClasses` will import the whole classpath.
This is not true anymore since release 1.0, because we changed the default behavior to import the package of the annotated test class.
The user guide now reflects this correctly and also mentions the option `wholeClasspath` to restore the old behavior.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>