extend CI to run tests with JDK 21
This was previously blocked by our Gradle version being too old.
Now that we can run the tests with JDK 21 it shows though,
that the `Annotation` `toString()` behavior has changed again.
Unfortunately, this always either causes complexity on the production or test side.
If we want to test the `AnnotationProxy.toString()` behavior,
then either we have to add a workaround to the test code,
or we have to adjust the production code to produce the equivalent `toString()` value.
I opted for the latter in hopes that they will finally stop changing the `toString()` style every couple of versions.

Signed-off-by: Peter Gafert <peter.gafert@archunit.org>