Ignore sun packages when importing full classpath
resolves #1446

The CI build has recently often failed on `ubuntu` and `macos` with Java 8,
due to `java.lang.OutOfMemoryError: GC overhead limit exceeded`.

This is due to `ClassFileImporterSlowTest` importing the full classpath in some tests,
which gives more than 20k classes on Java 8,
which contains 7k classes in `com.sun` and 4k classes in `sun` packages
that can easily be ignored.

Signed-off-by: Manfred Hanke <Manfred.Hanke@tngtech.com>